pyproject.toml
[tool.poetry]
name = "ClydeWhisper"
version = "0.1.0"
description = ""
authors = ["Ophelia <ophelia.pei@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
djangorestframework = "^3.14.0"
openai = "^0.27.8"
python-dotenv = "^1.0.0"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
